feat: use vet domains url to fetch avatar
