+++ packages/vechain-kit/src/hooks/api/vetDomains/index.ts
@@ -8,3 +8,4 @@ export * from './useGetTextRecords';
 export * from './useUpdateTextRecord';
 export * from './useGetResolverAddress';
 export * from './useGetAvatarOfAddress';
+export * from './useGetAvatarLegacy';
